Add ignore favicon functionality and es6 compatibility with babel build
