+++ .babelrc
@@ -1,10 +1,5 @@
 {
   "presets": [
-      "@babel/preset-env",
-      {
-        "targets": {
-          "node": "current"
-        }
-      }
+      "@babel/preset-env"
     ]
 }
