+++ .gitignore
@@ -6,6 +6,7 @@ coverage
 
 # production
 build
+dist
 
 # misc
 .DS_Store
