+++ lib/utils/ignoreFavicon.js
@@ -1,0 +1,7 @@
+export const ignoreFavicon = (req, res, next) => {
+	if (req.originalUrl === '/favicon.ico') {
+		res.status(204).json({nope: true});
+	} else {
+		next();
+	}
+};
