+++ public/index.html
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html lang="en" dir="ltr">
+  <head>
+    <meta charset="utf-8">
+    <title>Quick Connect</title>
+  </head>
+  <body>
+
+  </body>
+</html>
